feat(StatCalculator): 결과 계산 클래스 추가
- 수익률 계산 함수 이동
- 등수 계산 함수 이동
- 맞춘 숫자 개수에 따라 등수를 반환하는 함수 분리
