+++ src/main/kotlin/lotto/model/Lotto.kt
@@ -1,7 +1,6 @@
 package lotto.model
 
 import lotto.constants.ErrorMessage
-import lotto.utils.Validation
 
 class Lotto(private val numbers: List<Int>) {
     init {
@@ -10,20 +9,4 @@ class Lotto(private val numbers: List<Int>) {
     }
 
     fun getNumbers() = numbers
-
-    fun getRank(winningLotto: Set<Int>, bonusNumber: Int): Rank {
-        var count = 0
-        var matchBonusNumber = false
-        numbers.forEach { num ->
-            if(winningLotto.contains(num)) count++
-            if(num == bonusNumber) matchBonusNumber = true
-        }
-        return when(count) {
-            3 -> Rank.FIFTH
-            4 -> Rank.FOURTH
-            5 -> if(matchBonusNumber) Rank.SECOND else Rank.THIRD
-            6 -> Rank.FIRST
-            else -> Rank.NONE
-        }
-    }
 }
